mv number_count and fused_moe op to phi, fix some fused_moe bug
