+++ paddle/fluid/operators/fused/CMakeLists.txt
@@ -23,7 +23,6 @@ register_operators(
   fused_transformer_op
   fused_feedforward_op
   fused_multi_transformer_op
-  fused_moe_op
   fused_multi_transformer_int8_op
   fused_bias_dropout_residual_layer_norm_op
   resnet_unit_op
@@ -121,7 +120,6 @@ if(WITH_GPU OR WITH_ROCM)
     # fused_attention_op
     op_library(fused_attention_op)
     op_library(fused_multi_transformer_op)
-    op_library(fused_moe_op)
     op_library(fused_multi_transformer_int8_op)
     op_library(fused_bias_dropout_residual_layer_norm_op)
   endif()
