+++ paddle/fluid/pybind/op_function_generator.h
@@ -71,16 +71,6 @@ std::map<std::string, std::set<std::string>> op_ins_map = {
       "FFN1Bias",
       "FFN2Weight",
       "FFN2Bias"}},
-    {"fused_moe",
-     {"X",
-      "GateWeight",
-      "GateBias",
-      "LnScale",
-      "LnBias",
-      "ExpertsWeight1",
-      "ExpertsBias1",
-      "ExpertsWeight2",
-      "ExpertsBias2"}},
     {"fused_multi_transformer_int8",
      {"X",           "LnScale",           "LnBias",       "QKVW",
       "QKVBias",     "CacheKV",           "TimeStep",     "SrcMask",
@@ -345,7 +335,6 @@ std::map<std::string, std::set<std::string>> op_outs_map = {
       "Beta2PowOut",
       "MasterParamOut"}},
     {"fused_multi_transformer", {"CacheKVOut", "Out"}},
-    {"fused_moe", {"Out"}},
     {"fused_multi_transformer_int8", {"CacheKVOut", "Out"}},
     {"resnet_basic_block",
      {"Y",         "Conv1",     "SavedMean1", "SavedInvstd1", "Mean1Out",
