+++ paddle/phi/api/yaml/ops.yaml
@@ -217,3 +217,21 @@
   kernel :
     func : beam_search_softmax
     data_type : logits
+
+- op : number_count_v2
+  args : (Tensor numbers, int upper_range)
+  output : Tensor(out)
+  infer_meta :
+    func : NumberCountInferMeta
+  kernel :
+    func : number_count_v2
+    data_type : numbers
+
+- op : fused_moe_kernel
+  args : (Tensor x, Tensor gate_weight, Tensor gate_bias, Tensor ln_scale, Tensor ln_bias, Tensor[] experts_weight1, Tensor[] experts_bias1, Tensor[] experts_weight2, Tensor[] experts_bias2, bool pre_layer_norm, float ln_epsilon, int topk, int mp_size, int mp_rank, int num_expert, int world_size, int moe_ring_id, bool approximate)
+  output : Tensor(out)
+  infer_meta :
+    func : FusedMoeInferMeta
+  kernel :
+    func : fused_moe_kernel
+    data_type : x
